Revert https://src.chromium.org/viewvc/chrome?view=rev&revision=231330 This is a speculative revert to see if it's the cause of hanging renderers.
BUG=313112

Review URL: https://codereview.chromium.org/57783006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@232819 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 753bb25fc70b560fb3539e2b63843a193ccb295a

+++ base/pickle.cc
@@ -170,15 +170,15 @@ Pickle::Pickle(int header_size)
   header_->payload_size = 0;
 }
 
-Pickle::Pickle(const char* data, size_t data_len)
+Pickle::Pickle(const char* data, int data_len)
     : header_(reinterpret_cast<Header*>(const_cast<char*>(data))),
       header_size_(0),
       capacity_after_header_(kCapacityReadOnly),
       write_offset_(0) {
-  if (data_len >= sizeof(Header))
+  if (data_len >= static_cast<int>(sizeof(Header)))
     header_size_ = data_len - header_->payload_size;
 
-  if (header_size_ > data_len)
+  if (header_size_ > static_cast<unsigned int>(data_len))
     header_size_ = 0;
 
   if (header_size_ != AlignInt(header_size_, sizeof(uint32)))
